create class for convert_connectivity_nodes_cim16, fix old converter test
